Fixed issue with all minecarts using new physics which broke a lot of redstone contraptions. This version v1.0 makes it so player minecarts use new physics and the speeds can be changed and other minecarts use the old physics and do 8b/s. Working on making that changeble too.
How the fix works is that it checks if a minecart is ridable, if it is it tells the minecart to use new physics and if not use the old one. Because the experimental status is global variable we have to make a ThreadLocal and put the information here so the global variable dosent override the minecarts we want to use old physics.